# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SEN Migration Dashboard API
  description: |
    REST API for the SEN Migration Platform.
    Provides migration status, object progress, rule analysis,
    reconciliation, test scenario management, forensic extraction,
    process mining, data export, and migration planning.
  version: 1.0.0
  contact:
    name: SEN Team
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:4004
    description: Local development
  - url: https://{host}
    description: Production
    variables:
      host:
        default: sapconnect.example.com
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/export/forensic/json:
    get:
      tags: [Export]
      summary: Download forensic report (JSON)
      operationId: exportForensicJson
      responses:
        '200':
          description: Forensic report as downloadable JSON
          content:
            application/json:
              schema:
                type: object
        '404':
          description: No report available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/export/forensic/markdown:
    get:
      tags: [Export]
      summary: Download forensic report (Markdown)
      operationId: exportForensicMarkdown
      responses:
        '200':
          description: Forensic report as downloadable Markdown
          content:
            text/markdown:
              schema:
                type: string
        '404':
          description: No report available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/export/event-log/csv:
    get:
      tags: [Export]
      summary: Download event log (CSV)
      operationId: exportEventLogCsv
      responses:
        '200':
          description: Event log as downloadable CSV
          content:
            text/csv:
              schema:
                type: string
        '404':
          description: No event log available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/export/event-log/xes:
    get:
      tags: [Export]
      summary: Download event log (XES)
      operationId: exportEventLogXes
      responses:
        '200':
          description: Event log as downloadable XES (XML)
          content:
            application/xml:
              schema:
                type: string
        '404':
          description: No event log available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/export/event-log/json:
    get:
      tags: [Export]
      summary: Download event log (JSON)
      operationId: exportEventLogJson
      responses:
        '200':
          description: Event log as downloadable JSON
          content:
            application/json:
              schema:
                type: object
        '404':
          description: No event log available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/export/migration/plan/json:
    get:
      tags: [Export]
      summary: Download migration plan (JSON)
      operationId: exportMigrationPlanJson
      responses:
        '200':
          description: Migration plan as downloadable JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationPlan'
        '404':
          description: No plan available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Signavio endpoints ──────────────────────────────────────
...
  /api/cloud/sf/entities/{entitySet}:
    get:
      tags: [CloudSF]
      summary: Query SuccessFactors entities
      operationId: querySFEntities
      parameters:
        - name: entitySet
          in: path
          required: true
          schema:
            type: string
          example: PerPerson
      responses:
        '200':
          description: Entity query results
          content:
            application/json:
              schema:
                type: object
                properties:
                  entitySet:
                    type: string
                  results:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
        '404':
          description: Entity set not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [CloudSF]
      summary: Create SuccessFactors entity
      operationId: createSFEntity
      parameters:
        - name: entitySet
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Entity created
          content:
            application/json:
              schema:
                type: object

  # ── Cloud Ariba endpoints ──────────────────────────────────
...
  /api/cloud/concur/users:
    post:
      tags: [CloudConcur]
      summary: Provision Concur user
      operationId: provisionConcurUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
      responses:
        '201':
          description: User provisioned
          content:
            application/json:
              schema:
                type: object

  # ── Cloud SAC endpoints ────────────────────────────────────
...
  /api/cloud/sac/models/{id}/import:
    post:
      tags: [CloudSAC]
      summary: Import data into SAC model
      operationId: importSACData
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  modelId:
                    type: string
                  importStatus:
                    type: string
                  rowsImported:
                    type: integer
                  importedAt:
                    type: string
                    format: date-time
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Required when API_KEY environment variable is set.
        Pass via X-API-Key header or apiKey query parameter.
        Exempt paths: /health, /ready, /metrics
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
...
    MigrationPlan:
      type: object
      properties:
        generatedAt:
          type: string
          format: date-time
        scope:
          type: object
          properties:
            activeModules:
              type: array
              items:
                type: string
            totalObjects:
              type: integer
            totalEstimatedRecords:
              type: integer
        objects:
          type: array
          items:
            type: object
        executionPlan:
          type: object
          properties:
            waves:
              type: array
              items:
                type: object
            totalWaves:
              type: integer
        effort:
          type: object
        risks:
          type: array
          items:
            type: object
        recommendations:
          type: array
          items:
            type: object